+++ src/components/PageHeader/PageHeader.tsx
@@ -1,66 +1,113 @@
-import { styled } from "@mui/material/styles";
+import React from "react";
+import { styled } from "@mui/system";
+import avatar from "src/images/svg/avatar.svg";
+import logo from "src/images/svg/Logo.svg";
 
-import AppBar from "@mui/material/AppBar";
-import Box from "@mui/material/Box";
-import Toolbar from "@mui/material/Toolbar";
-import Container from "@mui/material/Container";
-import Avatar from "@mui/material/Avatar";
-import ListItem from "@mui/material/ListItem";
-import { Typography } from "@mui/material";
-
-type Props = {
-  name: string;
-};
-
-const CustomAppBar = styled(AppBar)({
-  "& .MuiContainer-root": {
-    backgroundColor: " #051424",
-    display: "flex",
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: " center",
-    position: "relative",
-    width: "95%",
-    margin: "0 0 0 5%",
-  },
-  "& .MuiPaper-root": {
-    backgroundColor: " #051424",
-    fontSize: 14,
-  },
-  "& .MuiAvatar-root": {
-    border: "3px solid #164370",
-  },
-  "& .MuiTypography-root": {
-    color: "#748BA0",
-    fontSize: 14,
-    margin: "0 0 0 10px",
-  },
-});
-const styles = {
+const PageHeaderBlock = styled("div")({
+  width: "100%",
+  height: "7%",
+  background: "#081A2C",
   display: "flex",
-  flexDirection: "row",
+  justifyContent: "space-between",
   alignItems: "center",
-};
+});
 
-const PageHeader = ({ name }: Props) => {
+const LogoPosition = styled("div")({
+  width: "10%",
+  height: "100%",
+  background: "#081A2C",
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "center",
+  borderRight: "1px solid #01070E",
+});
+
+const LogoImage = styled("img")({});
+
+const HeaderItems = styled("div")({
+  width: "90%",
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "space-between",
+});
+
+const HeaderLeft = styled("div")({
+  width: "50%",
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "space-evenly",
+  padding: "0 1%",
+});
+
+const HeaderRight = styled("div")({
+  width: "17%",
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "space-evenly",
+  padding: "0 3%",
+});
+
+const UserGreeting = styled("h2")({
+  fontWeight: "400",
+  fontSize: "14px",
+  color: "#748BA0",
+  marginRight: "5%",
+});
+
+const BlackBlock = styled("div")({
+  width: "170px",
+  height: "40px",
+  background: "black",
+  fontWeight: "400",
+  fontSize: "14px",
+  color: "#748BA0",
+  display: "flex",
+  justifyContent: "center",
+  alignItems: "center",
+  marginRight: "5%",
+});
+
+const SmallBlackBlock = styled("div")({
+  width: "114px",
+  height: "28px",
+  background: "black",
+  fontWeight: "400",
+  fontSize: "12px",
+  color: "#215FFF",
+  display: "flex",
+  justifyContent: "center",
+  alignItems: "center",
+});
+
+const Avatar = styled("img")({});
+
+const NameUser = styled("h2")({
+  fontWeight: "400",
+  fontSize: "14px",
+  color: "#748BA0",
+});
+
+const PageHeader = () => {
   return (
-    <CustomAppBar position="static">
-      <AppBar position="static">
-        <Container maxWidth="xl">
-          <Typography>Hello, {name}, Welcome to Spoon</Typography>
-          <Toolbar disableGutters>
-            <Box sx={styles}>/menu and date picker/</Box>
-          </Toolbar>
-          <Toolbar disableGutters>
-            <Box sx={styles}>
-              <ListItem>/last login/</ListItem>
-              <Avatar alt="Remy Sharp" src="/static/images/avatar/2.jpg" />
-              <ListItem>{name}</ListItem>
-            </Box>
-          </Toolbar>
-        </Container>
-      </AppBar>
-    </CustomAppBar>
+    <PageHeaderBlock>
+      <LogoPosition>
+        <LogoImage src={logo} alt="camera" />
+      </LogoPosition>
+
+      <HeaderItems>
+        <HeaderLeft>
+          <UserGreeting>Hello Magrot, Welcome to Spoons</UserGreeting>
+          <BlackBlock>Restaurant Open</BlackBlock>
+          <BlackBlock>24 September 2022</BlackBlock>
+        </HeaderLeft>
+
+        <HeaderRight>
+          <SmallBlackBlock>Last login: 13:22</SmallBlackBlock>
+          <Avatar src={avatar} />
+          <NameUser>Margot Crouch</NameUser>
+        </HeaderRight>
+      </HeaderItems>
+    </PageHeaderBlock>
   );
 };
