+++ src/components/PageSidebar/PageSidebar.tsx
@@ -3,7 +3,7 @@ import { styled } from "@mui/system";
 import TopSidebar from "./TopSidebar";
 import BottomSidebar from "./BottomSidebar";
 
-const NormalSidebar = styled("div")({
+const PageSidebarBlock = styled("div")({
   width: "10%",
   height: "100%",
   "@media (max-width:1700px)": {
@@ -14,10 +14,10 @@ const NormalSidebar = styled("div")({
 });
 const PageSidebar = () => {
   return (
-    <NormalSidebar>
-      <TopSidebar></TopSidebar>
-      <BottomSidebar></BottomSidebar>
-    </NormalSidebar>
+    <PageSidebarBlock>
+      <TopSidebar />
+      <BottomSidebar />
+    </PageSidebarBlock>
   );
 };
